Fixing ConfigProvider, day 2
