+++ src/components/ConfigProvider/ConfigProvider.tsx
@@ -1,5 +1,6 @@
-import React, { useContext } from "react";
-import {ThemeContext, useAppearance} from "../../hooks/themes";
+import React, { useContext, useState } from "react";
+import { ThemeContext } from "../../core/theme";
+import { useAppearance } from "../../hooks/themes";
 import "./ConfigProvider.css";
 
 export interface ConfigProviderProps {
@@ -10,15 +11,18 @@ export interface ConfigProviderProps {
 const ConfigProvider = (props: ConfigProviderProps) => {
   const appearance_default = useAppearance();
   const appearance = props.appearance || appearance_default;
-  const [theme, setTheme] = useContext(ThemeContext);
-  alert(setTheme);
-  // setTheme(appearance);
+  const [theme, setTheme] = useState("light");
+  const toggleAppearance = () => {
+    setTheme(
+      theme === "light" ? "dark" : "light"
+    )
+  }
   return (
-    <div
-      className={`configProvider ${appearance}`}
+    <ThemeContext.Provider
+      value = { { theme, setTheme } }
     >
       {props.children}
-    </div>
+    </ThemeContext.Provider>
   );
 };
